requirements.txt
flask>=3.0
flask-sock>=0.7
flask-cors>=4.0
google-cloud-firestore>=2.16
google-cloud-storage>=2.16
google-cloud-tasks>=2.16
google-genai>=1.0
google-auth>=2.29
Pillow>=10.0
mss>=9.0
playwright>=1.44
requests>=2.31
